whooo basic cell generation working

+++ src/forth/forth.h
@@ -13,7 +13,7 @@ typedef uint8_t generation_t;
 typedef uint32_t step_t;
 typedef uint8_t coord_t;
 typedef int8_t offset_t; /* [-1 .. 1] */
-typedef uint8_t energy_t;
+typedef uint16_t energy_t;
 
 
 typedef struct Forth {
